+++ tests/errors/nativeptr_err/nativeptr_err.c
@@ -1,32 +1,32 @@
-#error nativeptr_err.nativeptr_err:8:21 Expected a pointer type but was: unit
-#error nativeptr_err.nativeptr_err:9:27 Expected a pointer type but was: unativeint
-#error nativeptr_err.nativeptr_err:10:20 Expected a pointer type but was: obj
-#error nativeptr_err.nativeptr_err:11:25 Expected a pointer type but was: int list
-#error nativeptr_err.nativeptr_err:16:24 Expected two different pointer types but: one was unativeint, the other: voidptr
-#error nativeptr_err.nativeptr_err:17:27 Expected two different pointer types but: one was nativeptr<int>, the other: unativeint
-#error nativeptr_err.nativeptr_err:20:30 Expected two different pointer types but: one was nativeptr<'T>, the other: nativeptr<'T>
-#error nativeptr_err.nativeptr_err:23:7 This variable of owned type is not used.
-#error nativeptr_err.nativeptr_err:23:35 Expected two different pointer types but: one was voidptr, the other: Own<nativeptr<'T>>
-#error nativeptr_err.nativeptr_err:24:30 Expected two different pointer types but: one was Own<voidptr>, the other: nativeptr<'T>
-#error nativeptr_err.nativeptr_err:24:40 Expected a pointer type but was: Own<voidptr>
-#error nativeptr_err.nativeptr_err:28:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:29:15 Expected nativeptr, __outptr or voidptr type.
-#error nativeptr_err.nativeptr_err:32:15 Expected nativeptr, __outptr or voidptr type.
-#error nativeptr_err.nativeptr_err:34:3 Type mismatch: '__voidinptr' <> 'uint8'.
-#error nativeptr_err.nativeptr_err:35:3 Type mismatch: '__voidinptr' <> 'unit'.
-#error nativeptr_err.nativeptr_err:39:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:40:15 Expected __inptr, __voidinptr or __outptr type.
-#error nativeptr_err.nativeptr_err:44:9 Expected __inptr, __voidinptr or __outptr type.
-#error nativeptr_err.nativeptr_err:53:14 Type mismatch: 'nativeptr<int>' <> 'int'.
-#error nativeptr_err.nativeptr_err:54:14 Type mismatch: '__inptr<int>' <> 'nativeptr<int>'.
-#error nativeptr_err.nativeptr_err:55:14 Type mismatch: 'voidptr' <> 'nativeptr<int>'.
-#error nativeptr_err.nativeptr_err:64:11 Expected a variable.
-#error nativeptr_err.nativeptr_err:67:11 Expected a variable.
-#error nativeptr_err.nativeptr_err:71:22 Expected pointer type.
-#error nativeptr_err.nativeptr_err:73:24 Expected pointer type.
-#error nativeptr_err.nativeptr_err:79:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:80:20 Expected pointer type.
-#error nativeptr_err.nativeptr_err:81:25 Expected nativeptr but was __outptr.
+#error nativeptr_err.nativeptr_err:7:21 Expected a pointer type but was: unit
+#error nativeptr_err.nativeptr_err:8:27 Expected a pointer type but was: unativeint
+#error nativeptr_err.nativeptr_err:9:20 Expected a pointer type but was: obj
+#error nativeptr_err.nativeptr_err:10:25 Expected a pointer type but was: int list
+#error nativeptr_err.nativeptr_err:15:24 Expected two different pointer types but: one was unativeint, the other: voidptr
+#error nativeptr_err.nativeptr_err:16:27 Expected two different pointer types but: one was nativeptr<int>, the other: unativeint
+#error nativeptr_err.nativeptr_err:19:30 Expected two different pointer types but: one was nativeptr<'T>, the other: nativeptr<'T>
+#error nativeptr_err.nativeptr_err:22:7 This variable of owned type is not used.
+#error nativeptr_err.nativeptr_err:22:35 Expected two different pointer types but: one was voidptr, the other: Own<nativeptr<'T>>
+#error nativeptr_err.nativeptr_err:23:30 Expected two different pointer types but: one was Own<voidptr>, the other: nativeptr<'T>
+#error nativeptr_err.nativeptr_err:23:40 Expected a pointer type but was: Own<voidptr>
+#error nativeptr_err.nativeptr_err:27:15 This function misses some argument.
+#error nativeptr_err.nativeptr_err:28:15 Expected nativeptr, OutPtr or voidptr type.
+#error nativeptr_err.nativeptr_err:31:15 Expected nativeptr, OutPtr or voidptr type.
+#error nativeptr_err.nativeptr_err:33:3 Type mismatch: 'VoidInPtr' <> 'uint8'.
+#error nativeptr_err.nativeptr_err:34:3 Type mismatch: 'VoidInPtr' <> 'unit'.
+#error nativeptr_err.nativeptr_err:38:15 This function misses some argument.
+#error nativeptr_err.nativeptr_err:39:15 Expected InPtr, VoidInPtr or OutPtr type.
+#error nativeptr_err.nativeptr_err:43:9 Expected InPtr, VoidInPtr or OutPtr type.
+#error nativeptr_err.nativeptr_err:52:14 Type mismatch: 'nativeptr<int>' <> 'int'.
+#error nativeptr_err.nativeptr_err:53:14 Type mismatch: 'InPtr<int>' <> 'nativeptr<int>'.
+#error nativeptr_err.nativeptr_err:54:14 Type mismatch: 'voidptr' <> 'nativeptr<int>'.
+#error nativeptr_err.nativeptr_err:63:11 Expected a variable.
+#error nativeptr_err.nativeptr_err:66:11 Expected a variable.
+#error nativeptr_err.nativeptr_err:70:22 Expected pointer type.
+#error nativeptr_err.nativeptr_err:72:24 Expected pointer type.
+#error nativeptr_err.nativeptr_err:78:15 This function misses some argument.
+#error nativeptr_err.nativeptr_err:79:20 Expected pointer type.
+#error nativeptr_err.nativeptr_err:80:25 Expected nativeptr but was OutPtr.
+#error nativeptr_err.nativeptr_err:85:15 This function misses some argument.
 #error nativeptr_err.nativeptr_err:86:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:87:15 This function misses some argument.
-#error nativeptr_err.nativeptr_err:88:26 Expected nativeptr but was __inptr.
+#error nativeptr_err.nativeptr_err:87:26 Expected nativeptr but was InPtr.