+++ tests/primitives/native_type/native_type.fs
@@ -3,14 +3,14 @@ module rec native_type.Program
 // __nativeType<``ANY C TYPE``> is expanded to `ANY C TYPE` in C.
 // See also x_native_code.md in docs.
 
-module Ptr = Std.Ptr
+open Std.Ptr
 
 type FILE = __nativeType<FILE>
 
 // It can include spaces.
 type OpaqueStruct = __nativeType<``struct Opaque``>
 
-let strAsPtr (s: string) : __inptr<char> = __nativeFun ("string_as_ptr", s)
+let strAsPtr (s: string) : InPtr<char> = __nativeFun ("string_as_ptr", s)
 
 let fopen (filename: string) (mode: string) : nativeptr<FILE> =
   __nativeFun ("fopen", strAsPtr filename, strAsPtr mode)