+++ tests/primitives/nativeptr/nativeptr.fs
@@ -3,8 +3,8 @@ module rec nativeptr.Program
 // See x_native_code.md in docs.
 
 open Std.Own
-
-module Ptr = Std.Ptr
+open Std.Ptr
+open Std.Region
 
 let private memAlloc (count: uint) (size: uint) : voidptr =
   __nativeFun ("milone_region_alloc", count, size)
@@ -15,35 +15,35 @@ let private memSet (dest: voidptr) (value: uint8) (count: uint) =
 
   ()
 
-let private strcpy (dest: nativeptr<char>) (src: __inptr<char>) : nativeptr<char> = __nativeFun ("strcpy", dest, src)
+let private strcpy (dest: nativeptr<char>) (src: InPtr<char>) : nativeptr<char> = __nativeFun ("strcpy", dest, src)
 
 let private testBasic () =
   let buf = memAlloc 1u 8u
   memSet buf 255uy 8u
-  assert (Ptr.read (__nativeCast buf: __inptr<int>) = -1)
+  assert (Ptr.read (__nativeCast buf: InPtr<int>) = -1)
 
   // Conversion to int.
   assert (unativeint buf <> 0un)
 
 let private testVoidPtrAvailable () =
   let mutEnv: voidptr = __nativeCast 42un
-  let constEnv: __voidinptr = __nativeCast mutEnv
+  let constEnv: VoidInPtr = __nativeCast mutEnv
   assert (__nativeCast constEnv = 42un)
 
 let private testNullPtr () =
   let nullVoidPtr: voidptr = Ptr.nullPtr
   assert (__nativeCast nullVoidPtr = 0un)
 
-  let nullVoidInPtr: __voidinptr = Ptr.nullPtr
+  let nullVoidInPtr: VoidInPtr = Ptr.nullPtr
   assert (__nativeCast nullVoidInPtr = 0un)
 
   let nullNativePtr: nativeptr<float> = Ptr.nullPtr
   assert (__nativeCast nullNativePtr = 0un)
 
-  let nullInPtr: __inptr<float> = Ptr.nullPtr
+  let nullInPtr: InPtr<float> = Ptr.nullPtr
   assert (__nativeCast nullInPtr = 0un)
 
-  let nullOutPtr: __outptr<float> = Ptr.nullPtr
+  let nullOutPtr: OutPtr<float> = Ptr.nullPtr
   assert (__nativeCast nullOutPtr = 0un)
 
   let nullFunPtr: __nativeFun<unit, unit> = Ptr.nullPtr
@@ -61,12 +61,12 @@ let private testPtrInvalid () =
 let private testPtrCast () =
   // Upcast.
   let voidPtr: voidptr = Ptr.cast (Ptr.nullPtr: nativeptr<byte>)
-  let voidInPtr: __voidinptr = Ptr.cast (Ptr.nullPtr: __inptr<int>)
-  let intOutPtr: __outptr<int> = Ptr.cast (Ptr.nullPtr: nativeptr<int>)
+  let voidInPtr: VoidInPtr = Ptr.cast (Ptr.nullPtr: InPtr<int>)
+  let intOutPtr: OutPtr<int> = Ptr.cast (Ptr.nullPtr: nativeptr<int>)
 
   // Downcast.
   let intPtr: nativeptr<int> = Ptr.cast (Ptr.nullPtr: voidptr)
-  let objOutPtr: __outptr<obj> = Ptr.cast (Ptr.nullPtr: __inptr<string>)
+  let objOutPtr: OutPtr<obj> = Ptr.cast (Ptr.nullPtr: InPtr<string>)
   let funPtr: __nativeFun<obj * obj, obj> = Ptr.cast (Ptr.nullPtr: voidptr)
 
   // Own type can be cast.
@@ -78,22 +78,22 @@ let private testPtrCast () =
 
 let private testAsIn () =
   let mp: nativeptr<int> = Ptr.nullPtr
-  // __inptr<int>
+  // InPtr<int>
   let kp = Ptr.asIn mp
   assert (kp = Ptr.nullPtr)
 
   let mq: voidptr = Ptr.nullPtr
-  // __inptr<int>
+  // InPtr<int>
   let kq = Ptr.asIn mq
   assert (kq = Ptr.nullPtr)
 
 let private testAsNative () =
-  let kp: __inptr<int> = Ptr.nullPtr
+  let kp: InPtr<int> = Ptr.nullPtr
   // nativeptr<int>
   let mp = Ptr.asNative kp
   assert (mp = Ptr.nullPtr)
 
-  let kq: __voidinptr = Ptr.nullPtr
+  let kq: VoidInPtr = Ptr.nullPtr
   // voidptr
   let mq = Ptr.asNative kq
   assert (mq = Ptr.nullPtr)
@@ -110,8 +110,8 @@ let private testEquality () =
   let p: nativeptr<int> = __nativeCast 42un
   assert (p <> np)
 
-  let nq: __inptr<int> = Ptr.nullPtr
-  let q: __inptr<int> = __nativeCast 42un
+  let nq: InPtr<int> = Ptr.nullPtr
+  let q: InPtr<int> = __nativeCast 42un
   assert (q <> nq)
 
 let private sizeOfPointee (ptr: nativeptr<'T>) : int = sizeof<'T>
@@ -136,7 +136,7 @@ let private testSizeOf () =
 
 let private testPtrOf () =
   let x = 42
-  let p: __inptr<int> = &&x
+  let p: InPtr<int> = &&x
   assert (Ptr.read p = 42)
 
 let private testPtrSelect () =
@@ -154,7 +154,7 @@ let private testPtrRead () =
   assert (Ptr.read p.[0] = 1)
   assert (Ptr.read p.[4] = 16)
 
-  let q: __inptr<int> = Ptr.asIn p
+  let q: InPtr<int> = Ptr.asIn p
   assert (Ptr.read q = 1)
   assert (Ptr.read q.[2] = 4)
 
@@ -169,13 +169,13 @@ let private testPtrWrite () =
   Ptr.write p.[3] 39
   assert (Ptr.read p.[3] = 39)
 
-  let q: __outptr<int> = __nativeExpr "&data[7]"
+  let q: OutPtr<int> = __nativeExpr "&data[7]"
   Ptr.write q 77
   let q = Ptr.asNative q
   assert (Ptr.read q = 77)
 
-let private testPtrRegionAlloc () =
-  let p: __outptr<int> = Ptr.regionAlloc 2u
+let private testRegionAlloc () =
+  let p: OutPtr<int> = Region.alloc 2u
   Ptr.write p.[0] 42
   Ptr.write p.[1] 43
   // It's now initialized.
@@ -205,6 +205,6 @@ let main _ =
   testPtrSelect ()
   testPtrRead ()
   testPtrWrite ()
-  testPtrRegionAlloc ()
+  testRegionAlloc ()
   testPtrAddress ()
   0