+++ CHANGELOG.md
@@ -61,8 +61,8 @@ https://github.com/vain0x/milone-lang/commits/develop?after=689350a9d82133b68350
     - Support ownership check
     - Add several standard libraries built on top of it
 - Support more **pointer** types and primitives
-    - Add  `__outptr<'T>` and `__voidinptr<'T>` types
-    - Change `__constptr` to `__inptr`
+    - Add  `OutPtr<'T>` and `VoidInPtr<'T>` types
+    - Change `__constptr<'T>` to `InPtr<'T>`
     - Add ptr-of (`&&`) operator
     - Add `Ptr` module
         - `select`, `read`, `write`
@@ -73,7 +73,7 @@ https://github.com/vain0x/milone-lang/commits/develop?after=689350a9d82133b68350
     - Fix an issue that `string : bool -> string` didn't work at all
     - Add `__discriminant`
     - Add `sizeof<'T>`, remove `__sizeOfVal` primitive in favor of this
-    - Rename `inRegion` to `__inRegion`
+    - Rename `inRegion` to `Region.run`
     - `__nativeExpr` etc. can now embed some expressions and types
 - Support provisional manifest files (`milone_manifest`)
 
@@ -102,6 +102,7 @@ https://github.com/vain0x/milone-lang/commits/develop?after=689350a9d82133b68350
     - `StdTesting`
     - `Own`
     - `Ptr`
+    - `Region`
     - `IO`, `File`, `Path`, `OsString`
     - `Vector`, `ByteVector`, `Block`
 - Add functions: